refactor(web): shared PuzzleLifecycleAction + StatTile for admin pages
